split createNoteHeads function from abstract engraver
